feat : 로또 어플리케이션 작성

+++ src/test/kotlin/lotto/domain/vender/LottoTicketVendorTest.kt
@@ -8,7 +8,7 @@ internal class LottoTicketVendorTest {
 
     @ParameterizedTest
     @CsvSource(value = ["1000,1", "2000,2"])
-    fun `1000원에 1장씩 로또 티켓을 구매할수 있다`(price: Int, expect: Int) {
+    fun `1000원에 1장씩 로또 티켓을 구매할수 있다`(price: Long, expect: Int) {
         //given
         val vendor = LottoTicketVendor()
